requirements.txt
Django==1.11.1
pycrypto==2.6.1
